chore(SharedModule): reorganize file structure
