+++ src/app/shared/components/audio-timeline/audio-timeline.scss
