+++ src/app/shared/components/audio-timeline/index.ts
