+++ src/app/shared/components/content-header/content-header.scss
