+++ src/app/shared/components/icon-button/icon-button.scss
