+++ src/app/shared/components/icon/icon.scss
