+++ src/app/shared/components/icon/icon.spec.ts
