+++ src/app/shared/components/icon/icon.ts
