+++ src/app/shared/components/icon/index.ts
