+++ src/app/shared/components/loading-indicator/index.ts
