+++ src/app/shared/components/loading-indicator/loading-indicator.ts
