+++ src/app/shared/index.ts
@@ -1,0 +1,1 @@
+export { SharedModule } from './shared-module';
