+++ src/app/shared/pipes/format-integer/format-integer.spec.ts
