+++ src/app/shared/pipes/format-integer/format-integer.ts
