+++ src/app/shared/pipes/format-integer/index.ts
