+++ src/app/shared/pipes/format-time/format-time.ts
