+++ src/app/shared/shared-module.ts
@@ -1,12 +1,14 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
+// components
 import { AudioTimelineComponent } from './components/audio-timeline';
 import { ContentHeaderComponent } from './components/content-header';
 import { IconComponent } from './components/icon';
 import { IconButtonComponent } from './components/icon-button';
 import { LoadingIndicatorComponent } from './components/loading-indicator';
 
+// pipes
 import { FormatIntegerPipe } from './pipes/format-integer';
 import { FormatTimePipe } from './pipes/format-time';
