+++ src/index.html
@@ -7,14 +7,14 @@
 
   <title>SoundCloud • Angular2 NgRx</title>
 
-  <style><%= require('./shared/styles/styles.scss') %></style>
+  <style><%= require('./styles/styles.scss') %></style>
 
   <script src="https://use.typekit.net/wqt1ymx.js"></script>
   <script>try{Typekit.load({ async: true });}catch(e){}</script>
 </head>
 
 <body>
-  <%= require('raw-loader!./shared/assets/icons.svg') %>
+  <%= require('raw-loader!./assets/icons.svg') %>
 
   <app></app>
