+++ src/styles/_base.scss
@@ -1,5 +1,4 @@
-@import
-'settings',
+@import 'settings',
 'node_modules/minx/src/settings',
 'node_modules/minx/src/functions',
 'node_modules/minx/src/mixins';
