+++ src/styles/_settings.scss
