+++ src/styles/styles.scss
@@ -1,5 +1,4 @@
-@import
-'base',
+@import 'base',
 'node_modules/minx/src/reset',
 'node_modules/minx/src/elements',
 'elements',
