+++ src/tracklists/components/track-card.spec.ts
@@ -2,7 +2,7 @@ import { Component, Input, ViewChild } from '@angular/core';
 import { async, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { BehaviorSubject } from 'rxjs/BehaviorSubject';
-import { SharedModule } from 'src/shared';
+import { SharedModule } from 'src/app/shared';
 import { testUtils } from 'src/utils/test';
 import { createTrack } from '../models/track';
 import { TrackCardComponent } from './track-card';
