+++ src/users/components/user-card.spec.ts
@@ -1,7 +1,7 @@
 import { Component, ViewChild } from '@angular/core';
 import { async, TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
-import { SharedModule } from 'src/shared';
+import { SharedModule } from 'src/app/shared';
 import { UserRecord } from '../models/user';
 import { UserCardComponent } from './user-card';
